fixed code structure
fixed code readability
