+++ frontend/.gitignore
@@ -32,7 +32,7 @@ yarn-error.log*
 
 # env files (can opt-in for committing if needed)
 
-# .env*
+.env*
 
 # vercel
 .vercel
