+++ frontend/src/app/components/aboutComponent/about.css
