+++ frontend/src/app/components/dropdowns/dateDropdown.js
@@ -3,7 +3,7 @@ import { useEffect, useRef, useState } from 'react';
 import DateDropdown from '@/helper/dateButton';
 import DropdownContent from '@/helper/dateDropContent';
 import DropdownItem from '@/helper/dateItem';
-import styles from '@/app/styles/dateDrop.module.css';
+import styles from './dateDrop.module.css';
 const DropdownDate=({buttonText,content,selectedDate})=>{
      const [Dateopen,setDateOpen]=useState(false);
       const dropdownRef = useRef(null);
