+++ frontend/src/app/components/homeComponent/home-draw-1.css
