+++ frontend/src/app/components/homeComponent/home-draw-2.js
