+++ frontend/src/app/components/homeComponent/home-draw-3.css
