+++ frontend/src/app/components/mapComponent/comparisonStack.css
