+++ frontend/src/app/components/mapComponent/map-draw-1.js
