+++ frontend/src/app/components/mapComponent/map-draw-2.js
