+++ frontend/src/app/components/sidebar/weather-data.js
