+++ frontend/src/app/components/utils/button.js
@@ -1,4 +1,4 @@
-import '../styles/button.css';
+import './button.css';
 import Link from 'next/link';
 
 export default function Button(){
