+++ frontend/src/app/components/utils/dateUtils.js
@@ -1,0 +1,8 @@
+import {format} from 'date-fns';
+
+export function isToday(dateStr){
+const today=new Date();
+  const todayStr=format(today, 'MMMM d');
+  return dateStr===todayStr;
+}
+
